Modificación arquitectura Cliente / Servidor
- Issue #26
 Queda solucionada.
