--- ISO22_BC04/src/main/java/ISOFT4/bc04/dominio/Camino.java
@@ -1,73 +1,0 @@
-package ISOFT4.bc04.dominio;
-
-public class Camino {
-
-	private int idCamino;
-	private String nombreCamino;
-	private String tipoCamino;
-	private int dificultad;
-
-	/**
-	 * 
-	 * @param idCamino
-	 * @param nombreCamino
-	 * @param tipoCamino
-	 * @param dificultad
-	 */
-	public Camino(int idCamino, int nombreCamino, int tipoCamino, int dificultad) {
-		// TODO - implement Camino.Camino
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * 
-	 * @param nombreCamino
-	 * @param tipoCamino
-	 * @param dificultad
-	 */
-	public Camino AnadirCamino(int nombreCamino, int tipoCamino, int dificultad) {
-		// TODO - implement Camino.AnadirCamino
-		throw new UnsupportedOperationException();
-	}
-
-	public String getIdCamino() {
-		// TODO - implement Camino.getIdCamino
-		throw new UnsupportedOperationException();
-	}
-
-	public void setIdCamino() {
-		// TODO - implement Camino.setIdCamino
-		throw new UnsupportedOperationException();
-	}
-
-	public String getNombreCamino() {
-		// TODO - implement Camino.getNombreCamino
-		throw new UnsupportedOperationException();
-	}
-
-	public void setNombreCamino() {
-		// TODO - implement Camino.setNombreCamino
-		throw new UnsupportedOperationException();
-	}
-
-	public String getTipoCamino() {
-		// TODO - implement Camino.getTipoCamino
-		throw new UnsupportedOperationException();
-	}
-
-	public void setTipoCamino() {
-		// TODO - implement Camino.setTipoCamino
-		throw new UnsupportedOperationException();
-	}
-
-	public int getDificultad() {
-		// TODO - implement Camino.getDificultad
-		throw new UnsupportedOperationException();
-	}
-
-	public void setDificultad() {
-		// TODO - implement Camino.setDificultad
-		throw new UnsupportedOperationException();
-	}
-
-}
