--- ISO22_BC04/src/main/java/ISOFT4/bc04/dominio/Etapa.java
@@ -1,51 +1,0 @@
-package ISOFT4.bc04.dominio;
-
-import java.util.Date;
-
-public class Etapa {
-
-	private int idEtapa;
-	private Date fechaInicio;
-	private Date fechaFin;
-	private String nombreEtapa;
-	private String[] Rutas;
-
-	/**
-	 * 
-	 * @param idEtapa
-	 * @param fechaInicio
-	 * @param fechaFin
-	 * @param nombreEtapa
-	 * @param Rutas
-	 */
-	public Etapa(int idEtapa, int fechaInicio, int fechaFin, int nombreEtapa, int[] Rutas) {
-		// TODO - implement Etapa.Etapa
-		throw new UnsupportedOperationException();
-	}
-
-	public int getIdEtapa() {
-		// TODO - implement Etapa.getIdEtapa
-		throw new UnsupportedOperationException();
-	}
-
-	public Date getFechaInicio() {
-		// TODO - implement Etapa.getFechaInicio
-		throw new UnsupportedOperationException();
-	}
-
-	public Date getFechaFin() {
-		// TODO - implement Etapa.getFechaFin
-		throw new UnsupportedOperationException();
-	}
-
-	public String getNombreEtapa() {
-		// TODO - implement Etapa.getNombreEtapa
-		throw new UnsupportedOperationException();
-	}
-
-	public Ruta getRutas() {
-		// TODO - implement Etapa.getRutas
-		throw new UnsupportedOperationException();
-	}
-
-}
