--- ISO22_BC04/src/main/java/ISOFT4/bc04/dominio/Peregrino.java
@@ -1,113 +1,0 @@
-package ISOFT4.bc04.dominio;
-
-public class Peregrino {
-
-	private String nombre;
-	private String apellidos;
-	private int idPeregrino;
-	private int numTelf;
-	private String correo;
-	private String metodoPago;
-
-	/**
-	 * 
-	 * @param nombre
-	 * @param apellidos
-	 * @param numTelf
-	 * @param correo
-	 * @param metodoPago
-	 */
-	public Peregrino(int nombre, int apellidos, int numTelf, int correo, int metodoPago) {
-		// TODO - implement Peregrino.Peregrino
-		throw new UnsupportedOperationException();
-	}
-
-	public Etapa planificarRuta() {
-		// TODO - implement Peregrino.planificarRuta
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * 
-	 * @param Nombre
-	 */
-	public int crearIdPeregrino(int Nombre) {
-		// TODO - implement Peregrino.crearIdPeregrino
-		throw new UnsupportedOperationException();
-	}
-
-	public Ruta anadirRuta() {
-		// TODO - implement Peregrino.anadirRuta
-		throw new UnsupportedOperationException();
-	}
-
-	public Camino anadirCamino() {
-		// TODO - implement Peregrino.anadirCamino
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * 
-	 * @param metodoPago
-	 */
-	public String pago(int metodoPago) {
-		// TODO - implement Peregrino.pago
-		throw new UnsupportedOperationException();
-	}
-
-	public String getNombre() {
-		// TODO - implement Peregrino.getNombre
-		throw new UnsupportedOperationException();
-	}
-
-	public String setNombre() {
-		// TODO - implement Peregrino.setNombre
-		throw new UnsupportedOperationException();
-	}
-
-	public String getApellidos() {
-		// TODO - implement Peregrino.getApellidos
-		throw new UnsupportedOperationException();
-	}
-
-	public void setApellidos() {
-		// TODO - implement Peregrino.setApellidos
-		throw new UnsupportedOperationException();
-	}
-
-	public int getIdPeregrino() {
-		// TODO - implement Peregrino.getIdPeregrino
-		throw new UnsupportedOperationException();
-	}
-
-	public int getNumTelf() {
-		// TODO - implement Peregrino.getNumTelf
-		throw new UnsupportedOperationException();
-	}
-
-	public void setNumTelf() {
-		// TODO - implement Peregrino.setNumTelf
-		throw new UnsupportedOperationException();
-	}
-
-	public String getCorreo() {
-		// TODO - implement Peregrino.getCorreo
-		throw new UnsupportedOperationException();
-	}
-
-	public void setCorreo() {
-		// TODO - implement Peregrino.setCorreo
-		throw new UnsupportedOperationException();
-	}
-
-	public String getMetodoPago() {
-		// TODO - implement Peregrino.getMetodoPago
-		throw new UnsupportedOperationException();
-	}
-
-	public void setMetodoPago() {
-		// TODO - implement Peregrino.setMetodoPago
-		throw new UnsupportedOperationException();
-	}
-
-}
