--- ISO22_BC04/src/main/java/ISOFT4/bc04/dominio/Recurso.java
@@ -1,55 +1,0 @@
-package ISOFT4.bc04.dominio;
-
-public class Recurso {
-
-	private String nombreRecurso;
-	private String tipoRecuso;
-	private Ruta rutaLocalizacion;
-
-	/**
-	 * 
-	 * @param nombreRecurso
-	 * @param tipoRecurso
-	 * @param rutaLocalizacion
-	 */
-	public Recurso(int nombreRecurso, int tipoRecurso, int rutaLocalizacion) {
-		// TODO - implement Recurso.Recurso
-		throw new UnsupportedOperationException();
-	}
-
-	public void DisponibilidadRecurso() {
-		// TODO - implement Recurso.DisponibilidadRecurso
-		throw new UnsupportedOperationException();
-	}
-
-	public String getNombreRecurso() {
-		// TODO - implement Recurso.getNombreRecurso
-		throw new UnsupportedOperationException();
-	}
-
-	public void setNombreRecurso() {
-		// TODO - implement Recurso.setNombreRecurso
-		throw new UnsupportedOperationException();
-	}
-
-	public String getTipoRecurso() {
-		// TODO - implement Recurso.getTipoRecurso
-		throw new UnsupportedOperationException();
-	}
-
-	public void setTipoRecurso() {
-		// TODO - implement Recurso.setTipoRecurso
-		throw new UnsupportedOperationException();
-	}
-
-	public String getRutaLocalizacion() {
-		// TODO - implement Recurso.getRutaLocalizacion
-		throw new UnsupportedOperationException();
-	}
-
-	public void setRutaLocalizacion() {
-		// TODO - implement Recurso.setRutaLocalizacion
-		throw new UnsupportedOperationException();
-	}
-
-}
