--- ISO22_BC04/src/main/java/ISOFT4/bc04/dominio/Ruta.java
@@ -1,119 +1,0 @@
-package ISOFT4.bc04.dominio;
-
-public class Ruta {
-
-	private int IdRuta;
-	private String nombreRuta;
-	private double distanciaKM;
-	private String[] caminos;
-	private String[] recursos;
-	private String origen;
-	private String destino;
-
-	/**
-	 * 
-	 * @param idRuta
-	 * @param nombreRuta
-	 * @param distanciaKM
-	 * @param caminos
-	 * @param recursos
-	 * @param origen
-	 * @param destino
-	 */
-	public Ruta(int idRuta, int nombreRuta, int distanciaKM, int[] caminos, int[] recursos, int origen, int destino) {
-		// TODO - implement Ruta.Ruta
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * 
-	 * @param origen
-	 * @param destino
-	 */
-	public Ruta buscarRuta(int origen, int destino) {
-		// TODO - implement Ruta.buscarRuta
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * 
-	 * @param origen
-	 * @param destino
-	 * @param Caminos
-	 */
-	public Ruta crearRuta(int origen, int destino, int[] Caminos) {
-		// TODO - implement Ruta.crearRuta
-		throw new UnsupportedOperationException();
-	}
-
-	public void getIdRuta() {
-		// TODO - implement Ruta.getIdRuta
-		throw new UnsupportedOperationException();
-	}
-
-	public void setIdRuta() {
-		// TODO - implement Ruta.setIdRuta
-		throw new UnsupportedOperationException();
-	}
-
-	public void getNombreRuta() {
-		// TODO - implement Ruta.getNombreRuta
-		throw new UnsupportedOperationException();
-	}
-
-	public void setNombreRuta() {
-		// TODO - implement Ruta.setNombreRuta
-		throw new UnsupportedOperationException();
-	}
-
-	public void getDistanciaKM() {
-		// TODO - implement Ruta.getDistanciaKM
-		throw new UnsupportedOperationException();
-	}
-
-	public void setDistanciaKM() {
-		// TODO - implement Ruta.setDistanciaKM
-		throw new UnsupportedOperationException();
-	}
-
-	public void getOrigen() {
-		// TODO - implement Ruta.getOrigen
-		throw new UnsupportedOperationException();
-	}
-
-	public void setOrigen() {
-		// TODO - implement Ruta.setOrigen
-		throw new UnsupportedOperationException();
-	}
-
-	public void getDestino() {
-		// TODO - implement Ruta.getDestino
-		throw new UnsupportedOperationException();
-	}
-
-	public void setDestino() {
-		// TODO - implement Ruta.setDestino
-		throw new UnsupportedOperationException();
-	}
-
-	public void getCaminos() {
-		// TODO - implement Ruta.getCaminos
-		throw new UnsupportedOperationException();
-	}
-
-	public void setCaminos() {
-		// TODO - implement Ruta.setCaminos
-		throw new UnsupportedOperationException();
-	}
-
-	public void setRecursos() {
-		// TODO - implement Ruta.setRecursos
-		throw new UnsupportedOperationException();
-	}
-
-	public void getRecursos() {
-		// TODO - implement Ruta.getRecursos
-		throw new UnsupportedOperationException();
-	}
-
-}
