--- ISO22_BC04/src/main/java/ISOFT4/bc04/dominio/package-info.java
@@ -1,1 +1,0 @@
-package ISOFT4.bc04.dominio;
