+++ IT_1_PEREGRINOS/.settings/org.eclipse.jdt.core.prefs
