+++ IT_1_PEREGRINOS/pom.xml
@@ -2,12 +2,12 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>ISOFT4</groupId>
-  <artifactId>bc04</artifactId>
+  <groupId>ISO_BC044</groupId>
+  <artifactId>IT_1_PEREGRINOS</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <packaging>jar</packaging>
 
-  <name>bc04</name>
+  <name>IT_1_PEREGRINOS</name>
   <url>http://maven.apache.org</url>
 
   <properties>
