+++ IT_1_PEREGRINOS/src/main/java/ISO_BC044/IT_1_PEREGRINOS_DOMINIO/Camino.java
@@ -1,0 +1,32 @@
+package ISO_BC044.IT_1_PEREGRINOS_DOMINIO;
+public class Camino {
+
+	private int idCamino;
+	private String nombreCamino;
+	private String tipoCamino;
+	private int dificultad;
+
+	/**
+	 * 
+	 * @param idCamino
+	 * @param nombreCamino
+	 * @param tipoCamino
+	 * @param dificultad
+	 */
+	public void Camino(int idCamino, String nombreCamino, String tipoCamino, int dificultad) {
+		// TODO - implement Camino.Camino
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * 
+	 * @param nombreCamino
+	 * @param tipoCamino
+	 * @param dificultad
+	 */
+	public Camino anadirCamino(int nombreCamino, int tipoCamino, int dificultad) {
+		// TODO - implement Camino2.anadirCamino
+		throw new UnsupportedOperationException();
+	}
+
+}
