+++ IT_1_PEREGRINOS/src/main/java/ISO_BC044/IT_1_PEREGRINOS_DOMINIO/Peregrino.java
@@ -1,0 +1,50 @@
+package ISO_BC044.IT_1_PEREGRINOS_DOMINIO;
+
+public class Peregrino {
+
+	private String nombre;
+	private String apellidos;
+	private int idPeregrino;
+	private String numTelf;
+	private String email;
+	private String metodoPago;
+
+	/**
+	 * 
+	 * @param nombre
+	 * @param apellidos
+	 * @param idPeregrino
+	 * @param numTelf
+	 * @param email
+	 * @param metodoPago
+	 */
+	public Peregrino(String nombre, String apellidos, int idPeregrino, String numTelf, String email, String metodoPago) {
+		// TODO - implement Peregrino.Peregrino
+		throw new UnsupportedOperationException();
+	}
+
+	public Etapa planificarEtapa() {
+		// TODO - implement Peregrino.planificarEtapa
+		throw new UnsupportedOperationException();
+	}
+
+	public Ruta anadirRuta() {
+		// TODO - implement Peregrino.anadirRuta
+		throw new UnsupportedOperationException();
+	}
+
+	public Camino anadirCamino() {
+		// TODO - implement Peregrino.anadirCamino
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * 
+	 * @param metodoPago
+	 */
+	public String pago(int metodoPago) {
+		// TODO - implement Peregrino.pago
+		throw new UnsupportedOperationException();
+	}
+
+}
