+++ IT_1_PEREGRINOS/src/main/java/ISO_BC044/IT_1_PEREGRINOS_DOMINIO/Recurso.java
@@ -1,0 +1,25 @@
+package ISO_BC044.IT_1_PEREGRINOS_DOMINIO;
+
+public class Recurso {
+
+	private String nombreRecurso;
+	private String tipoRecurso;
+	private Ruta rutaLocalizacion;
+
+	/**
+	 * 
+	 * @param nombreRecurso
+	 * @param tipoRecurso
+	 * @param rutaLocalizacion
+	 */
+	public Recurso(String nombreRecurso, String tipoRecurso, Ruta rutaLocalizacion) {
+		// TODO - implement Recurso.Recurso
+		throw new UnsupportedOperationException();
+	}
+
+	public boolean disponibilidadRecurso() {
+		// TODO - implement Recurso.disponibilidadRecurso
+		throw new UnsupportedOperationException();
+	}
+
+}
