+++ IT_1_PEREGRINOS/src/main/java/ISO_BC044/IT_1_PEREGRINOS_DOMINIO/Ruta.java
@@ -1,0 +1,49 @@
+package ISO_BC044.IT_1_PEREGRINOS_DOMINIO;
+
+public class Ruta {
+
+	private int idRuta;
+	private String nombreRuta;
+	private double distanciaKM;
+	private String[] caminos;
+	private String[] recursos;
+	private String origen;
+	private String destino;
+
+	/**
+	 * 
+	 * @param idRuta
+	 * @param nombreRuta
+	 * @param distanciaKM
+	 * @param caminos
+	 * @param recursos
+	 * @param origen
+	 * @param destino
+	 */
+	public void Ruta(int idRuta, String nombreRuta, double distanciaKM, String[] caminos, String[] recursos, String origen, String destino) {
+		// TODO - implement Ruta
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * 
+	 * @param origen
+	 * @param destino
+	 */
+	public Ruta buscarRuta(String origen, String destino) {
+		// TODO - implement Ruta2.buscarRuta
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * 
+	 * @param origen
+	 * @param destino
+	 * @param Caminos
+	 */
+	public Ruta crearRuta(String origen, String destino, String[] Caminos) {
+		// TODO - implement Ruta2.crearRuta
+		throw new UnsupportedOperationException();
+	}
+
+}
