+++ IT_1_SERVIDOR/.gitignore
@@ -1,0 +1,2 @@
+/target
+.DS_Store
