+++ IT_1_SERVIDOR/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>ISO22_BC04</name>
+	<name>IT_1_SERVIDOR</name>
 	<comment></comment>
 	<projects>
 	</projects>
