+++ IT_1_SERVIDOR/.settings/org.maven.ide.eclipse.prefs
@@ -1,0 +1,8 @@
+#Tue Aug 23 02:56:03 PDT 2011
+activeProfiles=
+eclipse.preferences.version=1
+fullBuildGoals=process-test-resources
+resolveWorkspaceProjects=true
+resourceFilterGoals=process-resources resources\:testResources
+skipCompilerPlugin=true
+version=1
