+++ IT_1_SERVIDOR/src/main/java/ISO_BC044/IT_1_SERVIDOR_DOMINIO/Etapa.java
@@ -1,0 +1,26 @@
+package ISO_BC044.IT_1_SERVIDOR_DOMINIO;
+
+import java.util.Date;
+
+public class Etapa {
+
+	private int idEtapa;
+	private Date fechaInicio;
+	private Date fechaFin;
+	private String nombreEtapa;
+	private String[] Rutas;
+
+	/**
+	 * 
+	 * @param idEtapa
+	 * @param fechaInicio
+	 * @param fechaFin
+	 * @param nombreEtapa
+	 * @param Rutas
+	 */
+	public Etapa(int idEtapa, Date fechaInicio, Date fechaFin, String nombreEtapa, String[] Rutas) {
+		// TODO - implement Etapa.Etapa
+		throw new UnsupportedOperationException();
+	}
+
+}
