+++ IT_1_SERVIDOR/src/main/java/ISO_BC044/IT_1_SERVIDOR_PERSISTENCIA/Agente.java
@@ -1,0 +1,51 @@
+package ISO_BC044.IT_1_SERVIDOR_PERSISTENCIA;
+
+public class Agente {
+
+	public void Agente() {
+		// TODO - implement Agente.Agente
+		throw new UnsupportedOperationException();
+	}
+
+	public Agente getAgente() {
+		// TODO - implement Agente.getAgente
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * 
+	 * @param SQL
+	 */
+	public int insert(String SQL) {
+		// TODO - implement Agente.insert
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * 
+	 * @param SQL
+	 */
+	public String[] select(String SQL) {
+		// TODO - implement Agente.select
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * 
+	 * @param SQL
+	 */
+	public int update(String SQL) {
+		// TODO - implement Agente.update
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * 
+	 * @param SQL
+	 */
+	public int delete(String SQL) {
+		// TODO - implement Agente.delete
+		throw new UnsupportedOperationException();
+	}
+
+}
