+++ IT_1_SERVIDOR/src/main/java/ISO_BC044/IT_1_SERVIDOR_PRESENTACION/IU_Peregrino.java
@@ -1,0 +1,25 @@
+package ISO_BC044.IT_1_SERVIDOR_PRESENTACION;
+
+public class IU_Peregrino {
+
+	public void clickMostrarMenu() {
+		// TODO - implement IU_Peregrino2.clickMostrarMenu
+		throw new UnsupportedOperationException();
+	}
+
+	public void clickMostrarRecurso() {
+		// TODO - implement IU_Peregrino2.clickMostrarRecurso
+		throw new UnsupportedOperationException();
+	}
+
+	public void clickMostrarCamino() {
+		// TODO - implement IU_Peregrino2.clickMostrarCamino
+		throw new UnsupportedOperationException();
+	}
+
+	public void clickPlanificarEtapa() {
+		// TODO - implement IU_Peregrino2.clickPlanificarEtapa
+		throw new UnsupportedOperationException();
+	}
+
+}
